requirements.txt
openai==1.31
python-telegram-bot==20.4